refactor: move `createStyles` to parsers dir
