+++ src/utils/index.ts
@@ -1,7 +1,6 @@
 // Internal Utils
 export { addUnitIfNeeded } from './addUnitIfNeeded'
 export { createMediaQuery } from './createMediaQuery'
-export { createStyles } from './createStyles'
 export { getValue } from './getValue'
 export { isNumber } from './isNumber'
 export { merge } from './merge'
